+++ Tests/MirrorSharpOptionsWithXmlDocumentation.cs
@@ -4,7 +4,7 @@ using Microsoft.CodeAnalysis;
 
 namespace MirrorSharp.Tests {
     public static partial class MirrorSharpOptionsWithXmlDocumentation {
-        #if NETCOREAPP3_1
+        #if NETCOREAPP
         private static readonly string MscorlibReferenceAssemblyPath =
             AppDomain.CurrentDomain.BaseDirectory
                 + @"\ref-assemblies\System.Runtime.dll";